requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.6
django-filter==2.2.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
whitenoise==5.0.1
